utama/requirements.txt
fastapi
uvicorn
psycopg2
python-dotenv
Pillow
fuzzywuzzy
pandas
sentence-transformers
google-generativeai
pandas